+++ src/web/app/seed_loader.ts
@@ -5,7 +5,7 @@
 
 import { variations as proto } from '../../proto/generated/proto_bundle';
 import { SeedType, type ProcessingOptions } from '../../core/base_types';
-import { StudyListModel, StudyModel } from './models';
+import { StudyListModel, StudyModel } from './study_model';
 import { ProcessedStudy } from '../../core/study_processor';
 
 import * as url_utils from '../../core/url_utils';
